requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.6.15
charset-normalizer==2.0.12
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.5
gunicorn==20.1.0
Pillow==9.1.1
PyMySQL==1.0.2
python-dateutil==1.5
requests==2.28.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==3.3.0
djangorestframework~=3.13.1
pytest~=7.1.3
mysqlclient